+++ evalml/pipelines/pipeline_base.py
@@ -13,7 +13,7 @@ from .graphs import make_feature_importance_graph, make_pipeline_graph
 from evalml.exceptions import IllFormattedClassNameError
 from evalml.objectives import get_objective
 from evalml.problem_types import handle_problem_types
-from evalml.utils import Logger, classproperty
+from evalml.utils import Logger, classproperty, get_random_state
 
 logger = Logger()
 
@@ -25,30 +25,42 @@ class PipelineBase(ABC):
     @classmethod
     @abstractmethod
     def component_graph(cls):
+        """Returns list of components representing pipeline graph structure
+
+        Returns:
+            list(str/ComponentBase): list of ComponentBase objects or strings denotes graph structure of this pipeline
+        """
         return NotImplementedError("This pipeline must have `component_graph` as a class variable.")
 
     @property
     @classmethod
     @abstractmethod
     def supported_problem_types(cls):
+        """Returns a list of ProblemTypes that this pipeline supports
+
+        Returns:
+            list(str/ProblemType): list of ProblemType objects or strings that this pipeline supports
+        """
         return NotImplementedError("This pipeline must have `supported_problem_types` as a class variable.")
 
     custom_hyperparameters = None
 
-    def __init__(self, parameters, objective):
+    def __init__(self, parameters, objective, random_state=0):
         """Machine learning pipeline made out of transformers and a estimator.
 
         Required Class Variables:
             component_graph (list): List of components in order. Accepts strings or ComponentBase objects in the list
+
             supported_problem_types (list): List of problem types for this pipeline. Accepts strings or ProbemType enum in the list.
 
         Arguments:
             objective (ObjectiveBase): the objective to optimize
 
             parameters (dict): dictionary with component names as keys and dictionary of that component's parameters as values.
-                If `random_state`, `n_jobs`, or 'number_features' are provided as component parameters they will override the corresponding
-                value provided as arguments to the pipeline. An empty dictionary {} implies using all default values for component parameters.
+                 An empty dictionary {} implies using all default values for component parameters.
+            random_state (int, np.random.RandomState): The random seed/state. Defaults to 0.
         """
+        self.random_state = get_random_state(random_state)
         self.component_graph = [self._instantiate_component(c, parameters) for c in self.component_graph]
         self.supported_problem_types = [handle_problem_types(problem_type) for problem_type in self.supported_problem_types]
         self.objective = get_objective(objective)
@@ -115,7 +127,7 @@ class PipelineBase(ABC):
         component_name = component.name
         try:
             component_parameters = parameters.get(component_name, {})
-            new_component = component_class(**component_parameters)
+            new_component = component_class(**component_parameters, random_state=self.random_state)
         except (ValueError, TypeError) as e:
             err = "Error received when instantiating component {} with the following arguments {}".format(component_name, component_parameters)
             raise ValueError(err) from e