+++ evalml/tests/pipeline_tests/test_xgboost.py
@@ -23,6 +23,9 @@ def test_xg_init(X_y):
             'impute_strategy': 'median',
             'fill_value': None
         },
+        'One Hot Encoder': {
+            'top_n': 10
+        },
         'RF Classifier Select From Model': {
             "percent_features": 1.0,
             "number_features": len(X[0]),
@@ -37,12 +40,15 @@ def test_xg_init(X_y):
         }
     }
 
-    clf = XGBoostPipeline(objective=objective, parameters=parameters)
+    clf = XGBoostPipeline(objective=objective, parameters=parameters, random_state=1)
 
     expected_parameters = {
         'Simple Imputer': {
             'impute_strategy': 'median',
             'fill_value': None
+        },
+        'One Hot Encoder': {
+            'top_n': 10
         },
         'RF Classifier Select From Model': {
             'percent_features': 1.0,
@@ -57,6 +63,7 @@ def test_xg_init(X_y):
     }
 
     assert clf.parameters == expected_parameters
+    assert (clf.random_state.get_state()[0] == np.random.RandomState(1).get_state()[0])
 
 
 def test_xg_multi(X_y_multi):