+++ docs/source/changelog.rst
@@ -27,6 +27,7 @@ Changelog
         * Added automated dependency check PR :pr:`482`, :pr:`505`
         * Updated automated dependency check comment :pr:`497`
         * Have build_docs job use python executor, so that env vars are set properly :pr:`547`
+        * Run windows unit tests on PRs :pr:`557`
 .. warning::
 
     **Breaking Changes**